checkpoint Connected to AWS K8S cluster
Had to RENAME all contracts
