+++ artifacts2/DIDRegistry.ocean_poa_net_local.json
@@ -1,0 +1,191 @@
+{
+  "abi": [
+    {
+      "constant": false,
+      "inputs": [],
+      "name": "renounceOwnership",
+      "outputs": [],
+      "payable": false,
+      "stateMutability": "nonpayable",
+      "type": "function",
+      "signature": "0x715018a6"
+    },
+    {
+      "constant": true,
+      "inputs": [],
+      "name": "owner",
+      "outputs": [
+        {
+          "name": "",
+          "type": "address"
+        }
+      ],
+      "payable": false,
+      "stateMutability": "view",
+      "type": "function",
+      "signature": "0x8da5cb5b"
+    },
+    {
+      "constant": true,
+      "inputs": [],
+      "name": "isOwner",
+      "outputs": [
+        {
+          "name": "",
+          "type": "bool"
+        }
+      ],
+      "payable": false,
+      "stateMutability": "view",
+      "type": "function",
+      "signature": "0x8f32d59b"
+    },
+    {
+      "constant": false,
+      "inputs": [
+        {
+          "name": "newOwner",
+          "type": "address"
+        }
+      ],
+      "name": "transferOwnership",
+      "outputs": [],
+      "payable": false,
+      "stateMutability": "nonpayable",
+      "type": "function",
+      "signature": "0xf2fde38b"
+    },
+    {
+      "inputs": [],
+      "payable": false,
+      "stateMutability": "nonpayable",
+      "type": "constructor",
+      "signature": "constructor"
+    },
+    {
+      "anonymous": false,
+      "inputs": [
+        {
+          "indexed": true,
+          "name": "did",
+          "type": "bytes32"
+        },
+        {
+          "indexed": true,
+          "name": "owner",
+          "type": "address"
+        },
+        {
+          "indexed": true,
+          "name": "key",
+          "type": "bytes32"
+        },
+        {
+          "indexed": false,
+          "name": "value",
+          "type": "string"
+        },
+        {
+          "indexed": false,
+          "name": "valueType",
+          "type": "uint8"
+        },
+        {
+          "indexed": false,
+          "name": "updatedAt",
+          "type": "uint256"
+        }
+      ],
+      "name": "DIDAttributeRegistered",
+      "type": "event",
+      "signature": "0xfe303194f69c404a4ca19ca3d613a4bbcf419c764a463a930dd5686b5a6ba0f4"
+    },
+    {
+      "anonymous": false,
+      "inputs": [
+        {
+          "indexed": true,
+          "name": "previousOwner",
+          "type": "address"
+        },
+        {
+          "indexed": true,
+          "name": "newOwner",
+          "type": "address"
+        }
+      ],
+      "name": "OwnershipTransferred",
+      "type": "event",
+      "signature": "0x8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0"
+    },
+    {
+      "constant": false,
+      "inputs": [
+        {
+          "name": "_did",
+          "type": "bytes32"
+        },
+        {
+          "name": "_type",
+          "type": "uint8"
+        },
+        {
+          "name": "_key",
+          "type": "bytes32"
+        },
+        {
+          "name": "_value",
+          "type": "string"
+        }
+      ],
+      "name": "registerAttribute",
+      "outputs": [],
+      "payable": false,
+      "stateMutability": "nonpayable",
+      "type": "function",
+      "signature": "0x724ebe75"
+    },
+    {
+      "constant": true,
+      "inputs": [
+        {
+          "name": "_did",
+          "type": "bytes32"
+        }
+      ],
+      "name": "getUpdateAt",
+      "outputs": [
+        {
+          "name": "",
+          "type": "uint256"
+        }
+      ],
+      "payable": false,
+      "stateMutability": "view",
+      "type": "function",
+      "signature": "0xfc7bd95a"
+    },
+    {
+      "constant": true,
+      "inputs": [
+        {
+          "name": "_did",
+          "type": "bytes32"
+        }
+      ],
+      "name": "getOwner",
+      "outputs": [
+        {
+          "name": "",
+          "type": "address"
+        }
+      ],
+      "payable": false,
+      "stateMutability": "view",
+      "type": "function",
+      "signature": "0xdeb931a2"
+    }
+  ],
+  "bytecode": "0x608060405234801561001057600080fd5b5060008054600160a060020a0319163317808255604051600160a060020a039190911691907f8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0908290a3610522806100696000396000f3006080604052600436106100825763ffffffff7c0100000000000000000000000000000000000000000000000000000000600035041663715018a68114610087578063724ebe751461009e5780638da5cb5b146101075780638f32d59b14610138578063deb931a214610161578063f2fde38b14610179578063fc7bd95a1461019a575b600080fd5b34801561009357600080fd5b5061009c6101c4565b005b3480156100aa57600080fd5b50604080516020601f60643560048181013592830184900484028501840190955281845261009c9480359460ff602480359190911695604435953695608494930191819084018382808284375094975061022e9650505050505050565b34801561011357600080fd5b5061011c610409565b60408051600160a060020a039092168252519081900360200190f35b34801561014457600080fd5b5061014d610418565b604080519115158252519081900360200190f35b34801561016d57600080fd5b5061011c600435610429565b34801561018557600080fd5b5061009c600160a060020a0360043516610444565b3480156101a657600080fd5b506101b2600435610463565b60408051918252519081900360200190f35b6101cc610418565b15156101d757600080fd5b60008054604051600160a060020a03909116907f8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0908390a36000805473ffffffffffffffffffffffffffffffffffffffff19169055565b600084815260016020526040902054600160a060020a031680158061025b5750600160a060020a03811633145b15156102ee57604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152603060248201527f41747472696275746573206d757374206265207265676973746572656420627960448201527f2074686520444944206f776e6572732e00000000000000000000000000000000606482015290519081900360840190fd5b6040805180820182523380825243602080840182815260008b81526001808452908790209551865473ffffffffffffffffffffffffffffffffffffffff1916600160a060020a03909116178655905194019390935592518693919289927ffe303194f69c404a4ca19ca3d613a4bbcf419c764a463a930dd5686b5a6ba0f49288928b9291908190810184600381111561038357fe5b60ff168152602001838152602001828103825285818151815260200191508051906020019080838360005b838110156103c65781810151838201526020016103ae565b50505050905090810190601f1680156103f35780820380516001836020036101000a031916815260200191505b5094505050505060405180910390a45050505050565b600054600160a060020a031690565b600054600160a060020a0316331490565b600090815260016020526040902054600160a060020a031690565b61044c610418565b151561045757600080fd5b61046081610479565b50565b6000908152600160208190526040909120015490565b600160a060020a038116151561048e57600080fd5b60008054604051600160a060020a03808516939216917f8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e091a36000805473ffffffffffffffffffffffffffffffffffffffff1916600160a060020a03929092169190911790555600a165627a7a72305820d6e2cde4d4baf55d53f4dd73a0455ef7f7b326cce8c79f2d49a829a91dbfc54d0029",
+  "address": "0xbc79dC4785E9fd4B6357F184b41426C7E40EcF23"
+}
