+++ config_k8s_deployed.ini
@@ -1,5 +1,5 @@
 [keeper-contracts]
-keeper.url = http://ac9959fcade8d11e89c320e965e714bc-777187363.us-east-1.elb.amazonaws.com:8545
+keeper.url = http://52.1.94.55:8545
 keeper.path = artifacts
 
 secret_store.url = https://secret-store.dev-ocean.com/shadow/061299ac78ff49a19c1a284e7d3180c6131d88ce1fad45ca97e7f755acb694b1/7db43164e402dfddd1fb9bfc1f2ded608e2040962bb0f17af4728ba3c277772b1ba005f578293f9f3638e53bd6c365b34bc6e262f25686e7228a5becda4e197c01/2
@@ -10,15 +10,15 @@ parity.password = password
 
 [resources]
 azure.account.name = testocnfiles
-azure.account.key = DCNmb542DWbtkPf1lKz+WXii6Z50vScBvVJQMOy4XtG+bVIjHymbKm8iUZnSdlQlLRsrxlWaOwvAzbQmSm/oBw==
-azure.container = testfiles
-azure.resource_group = OceanProtocol
-azure.share.input = compute
-azure.share.output = output
-azure.location = westus
+azure.account.key =
+azure.container =
+azure.resource_group =
+azure.share.input =
+azure.share.output =
+azure.location =
 
 ;; These consitute part of the aquarius url which is used in setting the `api_url` in the `OceanContractsWrapper`
-aquarius.url = http://ac3195287e10911e89c320e965e714bc-1875844701.us-east-1.elb.amazonaws.com:5000
-brizo.url = http://a3c6e8416e40b11e88a360a98afc4587-44361392.us-east-1.elb.amazonaws.com:8030/
+aquarius.url = http://ac8b5e618ef0511e88a360a98afc4587-575519081.us-east-1.elb.amazonaws.com:5000
+brizo.url = http://ac8b8cc42ef0511e88a360a98afc4587-974193642.us-east-1.elb.amazonaws.com:8030
 
 storage.path = squid_py.db
